+++ NKPOS_V1/Services/DbServices/SetupCategoriesServices/SubCategoryServices/SubCategoryService.cs
@@ -19,7 +19,10 @@ namespace NKPOS_V1.Services.DbServices.SetupCategoriesServices.SubCategoryServic
                 {
                     return ResponseBuilder.CreateResponse(EnumStatusCode.NotFound, ResponseMessageUtils.CategoryNotFound);
                 }
-
+                if (await CheckSubCategoryDuplicate(model.SubCategoryName))
+                {
+                    return ResponseBuilder.CreateResponse(EnumStatusCode.Conflict, ResponseMessageUtils.DuplicateCategory(model.SubCategoryName));
+                }
                 await _context.SubCategories.AddAsync(model.ToEntity());
                 await _context.SaveChangesAsync();
 
@@ -31,22 +34,43 @@ namespace NKPOS_V1.Services.DbServices.SetupCategoriesServices.SubCategoryServic
             }
         }
 
+        private async Task<bool> CheckSubCategoryDuplicate(string subCategoryName)
+        {
+            bool isDuplicate = await _context.SubCategories.AnyAsync(c => c.SubCategoryName.ToLower() == subCategoryName.ToLower());
+            return isDuplicate;
+        }
+
         public async Task<SubCategoryListResponseModel> GetAllSubCategoriesAsync()
         {
             SubCategoryListResponseModel responseModel = new SubCategoryListResponseModel();
             try
             {
-                responseModel.DataLst = await _context.SubCategories
-                .Select(s => new SubCategoryResponseModel
-                {
-                    SubCategoryId = s.SubCategoryId,
-                    CategoryId = s.SubCategoryId,
-                    SubCategoryName = s.SubCategoryName,
-                    Description = s.Description,
-                    CreatedDate = s.CreatedDate,
-                    UpdatedDate = s.UpdatedDate,
-                })
-                .ToListAsync();
+                //responseModel.DataLst = await _context.SubCategories
+                //.Select(s => new SubCategoryResponseModel
+                //{
+                //    SubCategoryId = s.SubCategoryId,
+                //    CategoryId = s.SubCategoryId,
+                //    SubCategoryName = s.SubCategoryName,
+                //    Description = s.Description,
+                //    CreatedDate = s.CreatedDate,
+                //    UpdatedDate = s.UpdatedDate,
+                //})
+                //.ToListAsync();
+                responseModel.DataLst = await (from subcategory in _context.SubCategories.AsNoTracking()
+                                               join category in _context.Categories.AsNoTracking()
+                                               on subcategory.CategoryId equals category.CategoryId
+                                               orderby subcategory.SubCategoryId descending
+                                               select new SubCategoryResponseModel
+                                               {
+                                                   SubCategoryId = subcategory.SubCategoryId,
+                                                   SubCategoryName = subcategory.SubCategoryName,
+                                                   CategoryId = subcategory.CategoryId,
+                                                   CategoryName = category.CategoryName,
+                                                   Description = subcategory.Description,
+                                                   CreatedDate = subcategory.CreatedDate,
+                                                   UpdatedDate = subcategory.UpdatedDate
+                                               }).ToListAsync();
+
                 responseModel.baseResponseModel.RespCode = EnumStatusCode.Success;
                 responseModel.baseResponseModel.RespMessage = ResponseMessageUtils.Success;
                 return responseModel;
@@ -63,17 +87,32 @@ namespace NKPOS_V1.Services.DbServices.SetupCategoriesServices.SubCategoryServic
         {
             try
             {
-                return await _context.SubCategories
-                    .Where(s => s.SubCategoryId == subCategoryId && s.IsActive == true)
-                    .Select(s => new SubCategoryResponseModel
-                    {
-                        SubCategoryId = s.SubCategoryId,
-                        CategoryId = s.CategoryId,
-                        SubCategoryName = s.SubCategoryName,
-                        Description = s.Description,
-                        CreatedDate = s.CreatedDate,
-                        UpdatedDate = s.UpdatedDate,
-                    }).FirstOrDefaultAsync();
+                //return await _context.SubCategories
+                //    .Where(s => s.SubCategoryId == subCategoryId && s.IsActive == true)
+                //    .Select(s => new SubCategoryResponseModel
+                //    {
+                //        SubCategoryId = s.SubCategoryId,
+                //        CategoryId = s.CategoryId,
+                //        SubCategoryName = s.SubCategoryName,
+                //        Description = s.Description,
+                //        CreatedDate = s.CreatedDate,
+                //        UpdatedDate = s.UpdatedDate,
+                //    }).FirstOrDefaultAsync();
+
+                return await (from subcategory in _context.SubCategories.AsNoTracking()
+                              join category in _context.Categories.AsNoTracking()
+                              on subcategory.CategoryId equals category.CategoryId
+                              orderby subcategory.SubCategoryId descending
+                              select new SubCategoryResponseModel
+                              {
+                                  SubCategoryId = subcategory.SubCategoryId,
+                                  SubCategoryName = subcategory.SubCategoryName,
+                                  CategoryId = subcategory.CategoryId,
+                                  CategoryName = category.CategoryName,
+                                  Description = subcategory.Description,
+                                  CreatedDate = subcategory.CreatedDate,
+                                  UpdatedDate = subcategory.UpdatedDate
+                              }).FirstOrDefaultAsync();
             }
             catch (Exception ex)
             {
@@ -99,6 +138,11 @@ namespace NKPOS_V1.Services.DbServices.SetupCategoriesServices.SubCategoryServic
                     return ResponseBuilder.CreateResponse(EnumStatusCode.NotFound, ResponseMessageUtils.CategoryNotFound);
                 }
 
+                if (await CheckSubCategoryDuplicate(model.SubCategoryName))
+                {
+                    return ResponseBuilder.CreateResponse(EnumStatusCode.Conflict, ResponseMessageUtils.DuplicateSubCategory(model.SubCategoryName));
+                }
+
                 model.ToEntity(subCategory);
                 _context.SubCategories.Update(subCategory);
                 await _context.SaveChangesAsync();
@@ -122,6 +166,11 @@ namespace NKPOS_V1.Services.DbServices.SetupCategoriesServices.SubCategoryServic
                     return ResponseBuilder.CreateResponse(EnumStatusCode.NotFound, ResponseMessageUtils.SubCategoryNotFound);
                 }
 
+                if (await CheckSubCategoryMappedWithProduct(subCategoryId))
+                {
+                    return ResponseBuilder.CreateResponse(EnumStatusCode.BadRequest, ResponseMessageUtils.CannotDeleteSubCategory);
+                }
+
                 _context.SubCategories.Remove(subCategory);
                 await _context.SaveChangesAsync();
 
@@ -131,6 +180,16 @@ namespace NKPOS_V1.Services.DbServices.SetupCategoriesServices.SubCategoryServic
             {
                 return ResponseBuilder.CreateResponse(EnumStatusCode.InternalServerError, ex.Message);
             }
+        }
+
+        private async Task<bool> CheckSubCategoryMappedWithProduct(int subCategoryId)
+        {
+            return await (
+                from product in _context.Products
+                join subCategory in _context.SubCategories
+                    on product.SubCategoryId equals subCategory.SubCategoryId
+                select product
+            ).AnyAsync();
         }
     }
 }
